move test.py into sgrep lint
